refactor: move shared agent types to common/
`main/src/chat/agents/types.ts` was a pure types-and-constants module (`AgentConfig`, `BuiltinToolsKey`, `BUILTIN_TOOL_BUNDLES`, `BUILTIN_AGENT_IDS`, `DEFAULT_AGENT_ID`, …) consumed by all three packages — main, preload, and 11 renderer files. The renderer in particular reached across the package boundary with `../../../../../main/src/chat/agents/types`, mirroring a smell flagged in the #2184 review.

Move the file to `common/types/agents.ts`, where it sits alongside the other cross-package types (`common/types/cli.ts`, `common/types/toolhive-status.ts`), and switch every importer to the `@common/types/agents` alias that all three tsconfigs already wire up. No runtime behavior changes; the rename is detected by git so blame survives.

Resolves the deep-relative imports introduced by the new `SkillSelector` (#2184) and the pre-existing offenders in the agents feature.
